+++ src/backend/signaling-server/src/main/java/com/asyncgate/signaling_server/utility/S3Util.java
@@ -4,8 +4,8 @@ import com.amazonaws.SdkClientException;
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
-import com.asyncgate.user_server.exception.FailType;
-import com.asyncgate.user_server.exception.UserServerException;
+import com.asyncgate.signaling_server.exception.SignalingServerException;
+import com.asyncgate.signaling_server.exception.FailType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
@@ -39,19 +39,31 @@ public class S3Util {
 
             return bucketUrl + fileName;
         } catch (SdkClientException | IOException e) {
-            throw new UserServerException(FailType.UPLOAD_FILE_ERROR);
+            throw new SignalingServerException(FailType._UPLOAD_FILE_ERROR);
+        } catch (Exception e) {
+            throw new SignalingServerException(FailType._UNKNOWN_ERROR);
         }
     }
 
     // S3 파일 제거
     public void deleteFile(final String fileUrl) {
         try {
-            // 공백 제거
+            // fileName 추출
             String fileName = fileUrl.replace(bucketUrl, "");
 
+            // 파일 존재 여부 확인
+            boolean fileExists = amazonS3Client.doesObjectExist(bucketName, fileName);
+
+            if (!fileExists) {
+                throw new SignalingServerException(FailType._FILE_NOT_FOUND);
+            }
+
             amazonS3Client.deleteObject(bucketName, fileName);
+
         } catch (SdkClientException e) {
-            throw new UserServerException(FailType.DELETE_FILE_ERROR);
+            throw new SignalingServerException(FailType._DELETE_FILE_ERROR);
+        } catch (Exception e) {
+            throw new SignalingServerException(FailType._UNKNOWN_ERROR);
         }
     }
 }
